Допилил прогу №2
