+++ .env
@@ -1,3 +1,4 @@
 POSTGRES_USER=postgres
 POSTGRES_PASSWORD=Hanter678pqo
 POSTGRES_URL=TacrolimusDB
+DESTINATION_FILE_PATH=C:/santa_files/
