+++ src/main/java/com/tacrolimus/backend/config/APIConfig.java
@@ -9,9 +9,11 @@ public class APIConfig {
         String postgresUser = dotenv.get("POSTGRES_USER");
         String postgresPassword = dotenv.get("POSTGRES_PASSWORD");
         String postgresUrl = dotenv.get("POSTGRES_URL");
+        String destinationPath = dotenv.get("DESTINATION_FILE_PATH");
 
         System.setProperty("POSTGRES_USER", postgresUser);
         System.setProperty("POSTGRES_PASSWORD", postgresPassword);
         System.setProperty("POSTGRES_URL", postgresUrl);
+        System.setProperty("DESTINATION_FILE_PATH",destinationPath);
     }
 }
