--- src/main/java/com/tacrolimus/backend/dto/BaseFileInfoSaveDto.java
@@ -1,16 +1,0 @@
-package com.tacrolimus.backend.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@SuperBuilder
-public class BaseFileInfoSaveDto {
-    private String url;
-    private String name;
-    private int size;
-}
