--- src/main/java/com/tacrolimus/backend/dto/FileInfoUpdateDto.java
@@ -1,13 +1,0 @@
-package com.tacrolimus.backend.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.experimental.SuperBuilder;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@SuperBuilder
-public class FileInfoUpdateDto extends BaseFileInfoSaveDto{
-}
