+++ src/main/java/com/tacrolimus/backend/dto/RegistrationCreateDto.java
@@ -1,12 +1,10 @@
 package com.tacrolimus.backend.dto;
 
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 import lombok.experimental.SuperBuilder;
 
-@Getter
-@Setter
+@Data
+@EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor
 @SuperBuilder
 public class RegistrationCreateDto extends BaseRegistrationSaveDto{
