+++ src/main/java/com/tacrolimus/backend/dto/StatusReachedDto.java
@@ -1,6 +1,7 @@
 package com.tacrolimus.backend.dto;
 
 import com.tacrolimus.backend.enu.SantaPairStatusEnum;
+import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -8,8 +9,7 @@ import lombok.experimental.SuperBuilder;
 
 import java.util.UUID;
 
-@Getter
-@Setter
+@Data
 @NoArgsConstructor
 @SuperBuilder
 public class StatusReachedDto {
