+++ src/main/java/com/tacrolimus/backend/dto/StatusReadDto.java
@@ -1,13 +1,11 @@
 package com.tacrolimus.backend.dto;
 
 import com.tacrolimus.backend.enu.SantaPairStatusEnum;
-import lombok.Getter;
+import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 import lombok.experimental.SuperBuilder;
 
-@Getter
-@Setter
+@Data
 @NoArgsConstructor
 @SuperBuilder
 public class StatusReadDto {
